% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_SDC}
\alias{apply_SDC}
\title{Apply Statistical disclosure to numbers}
\usage{
apply_SDC(numbers, round_to = 10, issue_level = 5)
}
\arguments{
\item{numbers}{numbers to apply SDC to.}

\item{round_to}{number to round to}

\item{issue_level}{number, where numbers with fewer students have their value omitted (set to "Below issue_level").}
}
\value{
vector containing the numbers after performing SDC
}
\description{
Apply Statistical disclosure to numbers
}
