% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_frame_to_excel}
\alias{data_frame_to_excel}
\title{Write data frame to excel with title and source}
\usage{
data_frame_to_excel(
  data,
  title = "",
  source = "",
  filepath = paste0(getwd(), "/data_frame_info.xlsx")
)
}
\arguments{
\item{data}{data frame}

\item{title}{title for excel sheet}

\item{source}{source information of the data frame.}

\item{filepath}{path to save the excel file.}
}
\description{
Write data frame to excel with title and source
}
\examples{
if(F){
mtcars[1:10,] |> data_frame_to_excel(title = 'First 10 rows of mtcars',
source = 'Source: R', filepath = 'first_10_rows_mtcars.xlsx')
}
}
