% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_frame_to_workbook}
\alias{data_frame_to_workbook}
\title{Write data frame to openxlsx workbook}
\usage{
data_frame_to_workbook(
  append_to = NULL,
  data = NULL,
  sheet = "Data",
  title = "",
  source = ""
)
}
\arguments{
\item{append_to}{either NULL for a new workbook or a workbook to append the sheet to.}

\item{data}{data frame}

\item{sheet}{Sheet name}

\item{title}{title for excel sheet}

\item{source}{source information of the data frame.}
}
\value{
workbook
}
\description{
Write data frame to openxlsx workbook
}
