% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fake_data}
\alias{fake_data}
\title{Fake}
\format{
## `fake_data`
A data frame with 10,000 rows and 7 columns:
\describe{
  \item{FSM}{Does the student recieve Free School Meals (FSM).}
  \item{IDACI}{What IDACI quintie does the student belong to.}
  \item{Gender}{What Gender is the student.}
  \item{KS1}{KS1 maths outcome.}
  \item{KS2}{KS2 maths outcome.}
  \item{GCSE}{GCSE mathematics grade.}
  \item{A_level}{Does the student study maths and/or further maths at A level.}
}
}
\usage{
fake_data
}
\description{
A fake dataset of student information containing maths grades at KS1 - KS4 and some demographics.
}
\keyword{datasets}
