% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_functions.R
\name{filter_data}
\alias{filter_data}
\title{Filter the raw data}
\usage{
filter_data(data, filters, union = T)
}
\arguments{
\item{data}{data}

\item{filters}{filters}

\item{union}{union}
}
\description{
Filter the raw data
}
