% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey_helpers.R
\name{get_nodes}
\alias{get_nodes}
\alias{get_links}
\title{Extract nodes/links from a  Sankey diagram}
\usage{
get_nodes(p)

get_links(p)
}
\arguments{
\item{p}{sankey diagram object}
}
\description{
- `get_nodes()` returns the ordered named nodes of the Sankey diagram
}
\details{
- `get_links()`returns the ordered named links of the Sankey diagram.
}
