% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sankey_functions.R
\name{links_from_extract}
\alias{links_from_extract}
\title{Get node and links from excel file or extracted data}
\usage{
links_from_extract(filepath = NULL, extract = NULL, skip = 3)
}
\arguments{
\item{filepath}{filepath to the .xlsx file containing the sankey diagram information. Assume the first 3 rows are to be removed but this can be altered with the `skip` input.}

\item{extract}{data frame containing the sankey information, with column  names including Source, Target and Number of Students.}

\item{skip}{Number of rows to be skipped when loading the excel file.}
}
\value{
list containing nodes and links
}
\description{
Get node and links from excel file or extracted data
}
