% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LStools-package.R
\docType{package}
\name{LStools-package}
\alias{LStools}
\alias{LStools-package}
\title{LStools: Tools for exploring longitudinal (education) data (sankey diagrams)}
\description{
functions for creating sankey diagrams (build upon networkD3 package) for longitudinal (education) data together with other useful functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jakepowellcubg.github.io/LStools/}
  \item \url{https://jake-powell.github.io/LStools/}
}

}
\keyword{internal}
