% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_functions.R
\name{pathway_composition_tables}
\alias{pathway_composition_tables}
\title{A function to get the summary of the describers at each stage of the pipeline.}
\usage{
pathway_composition_tables(
  pathways,
  describers,
  stages,
  do_SDC = FALSE,
  round_to = 10,
  issue_level = 5
)
}
\arguments{
\item{pathways}{data frame containing pathways.}

\item{describers}{A character vector of column names in pathways which we want to describe the stages (such as FSM, Sex, IDACI, etc)}

\item{stages}{A character vector of column names in pathways we split by the columns given in describers. (columns such as KS1, KS2, etc)}

\item{do_SDC}{Flag (TRUE/FALSE) for whether statistical disclosure is performed.}

\item{round_to}{number to round to}

\item{issue_level}{number, where values with fewer students have their value omitted (set to "-").}
}
\description{
A function to get the summary of the describers at each stage of the pipeline.
}
