% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_functions.R
\name{pathway_remove_columns}
\alias{pathway_remove_columns}
\title{Remove columns from pathways}
\usage{
pathway_remove_columns(
  pathways,
  rm_columns = NA,
  keep_columns = NA,
  convert_na = T,
  count_column = "No_students"
)
}
\arguments{
\item{pathways}{data frame containing pathways.}

\item{rm_columns}{The columns names in pathways that are to be removed.}

\item{keep_columns}{The columns names in pathways that are to be kept.}

\item{convert_na}{Logical (TRUE/FALSE) for whether we convert NA.}

\item{count_column}{Default = 'No_students'. The name of the column containing the number of students.}
}
\value{
pathways with columns removed and merged counts.
}
\description{
A function to remove columns from the pathway tibble, merging now identical rows and summing their No_students.
}
