% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_functions.R
\name{pathway_stage_transition_by_describer}
\alias{pathway_stage_transition_by_describer}
\title{A function to get the summary of the describers at each stage of the pipeline.}
\usage{
pathway_stage_transition_by_describer(pathways, describers, stages)
}
\arguments{
\item{pathways}{data frame containing pathways.}

\item{describers}{A character vector of column names in pathways which we want to describe the stages (such as FSM, Sex, IDACI, etc)}

\item{stages}{A character vector of column names in pathways we split by the columns given in describers. (columns such as KS1, KS2, etc), note here that the order of stages need to be in chronological order.}
}
\description{
A function to get the summary of the describers at each stage of the pipeline.
}
