% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_functions.R
\name{pathway_summary_columns}
\alias{pathway_summary_columns}
\title{Composition of pathways by columns}
\usage{
pathway_summary_columns(pathways, convert_na = T)
}
\arguments{
\item{pathways}{data frame containing pathways.}

\item{convert_na}{Logical (TRUE/FALSE) for whether we convert NA.}
}
\description{
A function to get the summary of each column in pathways. I.e the Number of students in each stage/describer and how their split.
}
