% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_functions.R
\name{pathway_summary_describers_at_stages}
\alias{pathway_summary_describers_at_stages}
\title{A function to get the summary of the describers at each stage of the pipeline.}
\usage{
pathway_summary_describers_at_stages(pathways, describers, stages)
}
\arguments{
\item{pathways}{pathways}

\item{describers}{describers}

\item{stages}{stages}
}
\description{
A function to get the summary of the describers at each stage of the pipeline.
}
