% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_functions.R
\name{raw_to_pathway}
\alias{raw_to_pathway}
\title{Convert raw information (individual students) into pathways}
\usage{
raw_to_pathway(raw)
}
\arguments{
\item{raw}{raw data where each row is a unique student.}
}
\value{
pathways, the number of students for each unique pathway.
}
\description{
This function is to convert raw student information into flows of students through the pipeline.
 E.g. in raw each row is a unique student and each column contains information about the student such as grade or ethnicity.
 The function assumes each column is a variable we want to calculate flows for. Therefore, omit columns you don't need before using this function.
}
