% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove.factors}
\alias{remove.factors}
\title{Remove factors from a data frame.}
\usage{
remove.factors(df)
}
\arguments{
\item{df}{data frame}
}
\value{
data frame without factors
}
\description{
Remove factors from a data frame.
}
