% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sankey_functions.R
\name{sankey_add_column_names}
\alias{sankey_add_column_names}
\title{Add column labels to a Sankey diagram}
\usage{
sankey_add_column_names(
  p,
  column_titles,
  fontSize = p$x$options$fontSize,
  fontFamily = p$x$options$fontFamily
)
}
\arguments{
\item{p}{sankey diagram}

\item{column_titles}{column_titles}

\item{fontSize}{fontSize}

\item{fontFamily}{fontFamily}
}
\description{
Add column labels to a Sankey diagram
}
