% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey_helpers.R
\name{sankey_export_xlsx}
\alias{sankey_export_xlsx}
\title{Write sankey to excel}
\usage{
sankey_export_xlsx(
  p,
  title = "Sankey Information",
  source = "",
  filepath = paste0(getwd(), "/sankey_info.xlsx")
)
}
\arguments{
\item{p}{sankey diagram}

\item{title}{title for excel spreadsheet}

\item{source}{sources used for sankey diagram (NPD / HESA, etc)}

\item{filepath}{the file path to save the .xlsx file to.}
}
\description{
This function writes the sankey diagram information (flows) to an .xlsx file with title (title) and subtitle (source) ready for SDC checks / outputting.
}
