% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sankey_functions.R
\name{sankey_filtered}
\alias{sankey_filtered}
\title{A function that returns a sanket of the filtered data.}
\usage{
sankey_filtered(
  pathways,
  flow_columns,
  filters,
  union = F,
  FontFamily = NULL,
  FontSize = 7
)
}
\arguments{
\item{pathways}{pathways}

\item{flow_columns}{flow_columns}

\item{filters}{filters}

\item{union}{union}

\item{FontFamily}{colors}

\item{FontSize}{colors}
}
\value{
sankey diagram
}
\description{
A function that returns a sanket of the filtered data.
}
