% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sankey_functions.R
\name{sankey_from_extract}
\alias{sankey_from_extract}
\title{Create sankey diagram from excel file or extracted data}
\usage{
sankey_from_extract(
  filepath = NULL,
  extract = NULL,
  fontFamily = NULL,
  skip = 3,
  fontSize = 7
)
}
\arguments{
\item{filepath}{filepath to the .xlsx file containing the sankey diagram information. Assume the first 3 rows are to be removed but this can be altered with the `skip` input.}

\item{extract}{data frame containing the sankey information, with column  names including Source, Target and Number of Students.}

\item{fontFamily}{font family for the node text labels.}

\item{skip}{Number of rows to be skipped when loading the excel file.}

\item{fontSize}{numeric font size in pixels for the node text labels.}
}
\value{
sankey diagram
}
\description{
Create sankey diagram from excel file or extracted data
}
