% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sankey_functions.R
\name{sankey_trace}
\alias{sankey_trace}
\title{Create traced sankey diagram}
\usage{
sankey_trace(
  pathways,
  flow_columns,
  split_by,
  union = F,
  colors = c("red", "gray", "antiquewhite"),
  fontFamily = NULL,
  fontSize = 7
)
}
\arguments{
\item{pathways}{student pathways with number of students.}

\item{flow_columns}{Column names (or index) of the 'stages' in pathways that we want to create a sankey diagram of.}

\item{split_by}{a character vector describing the filters to split the data by. Of the format `XX: YY` where XX is the stage and YY the node, such as 'FSM: Yes'.}

\item{union}{Flag (TRUE/FALSE) for whether we want the union of filters (only across single stages)}

\item{colors}{a character vector of length three `c(colA, colB, colC)` where colA is the link colour for traced students, colB is the link colour for other students and colC is the node colour.}

\item{fontFamily}{font family for the node text labels.}

\item{fontSize}{numeric font size in pixels for the node text labels.}
}
\value{
traced sankey diagram
}
\description{
Create traced sankey diagram
}
