% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sankey_functions.R
\name{sankey_trace_groups}
\alias{sankey_trace_groups}
\title{Create sankey diagram where we show flow across multiple groups (using a single column)}
\usage{
sankey_trace_groups(
  pathways,
  flow_columns,
  group_column,
  fontFamily = NULL,
  fontSize = 7,
  do_SDC = FALSE,
  colors = viridis::viridis(10),
  ...
)
}
\arguments{
\item{pathways}{student pathways with number of students.}

\item{flow_columns}{Column names (or index) of the 'stages' in pathways that we want to create a sankey diagram of.}

\item{group_column}{the name (or index) of the group column}

\item{fontFamily}{font family for the node text labels.}

\item{fontSize}{numeric font size in pixels for the node text labels.}

\item{do_SDC}{Flag (TRUE/FALSE) for whether you want to do SDC via the `SDC_to_links()` function.}

\item{colors}{a character vector of length at least the number of groups to colour the edges in the sankey diagram}

\item{...}{arguments to `SDC_to_links()` or `networkD3::sankeyNetwork()`.}
}
\value{
sankey diagram
}
\description{
Create sankey diagram where we show flow across multiple groups (using a single column)
}
