% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sankey_helpers.R
\name{get_nodes}
\alias{get_nodes}
\alias{get_links}
\title{Sankey Helper functions}
\usage{
get_nodes(p)

get_links(p)
}
\arguments{
\item{p}{sankey diagram}
}
\description{
Helper functions that may be of use if you want to change the colour/text/hovertext of nodes and links in a sankey diagram.
}
\details{
- `get_nodes()` returns the ordered named nodes of the Sankey diagram

- `get_links()`returns the ordered named links of the Sankey diagram.
}
