% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sankey_functions.R
\name{pathway_to_links}
\alias{pathway_to_links}
\title{A function that converts pathways to sankey diagram information.}
\usage{
pathway_to_links(pathways, links_as_pos = TRUE, remove_na = T, ...)
}
\arguments{
\item{pathways}{pathways}

\item{links_as_pos}{links_as_pos}

\item{remove_na}{remove_na Flag (TRUE/FALSE) for whether NA values are removed.}

\item{...}{arguments to `SDC_to_links()`.}
}
\value{
links
}
\description{
Note that you need to remove "describer" columns prior to using this function. In other words each column in pathways is viewed as a stage (KS1, KS2, etc) in the sankey diagram. Hence the order of columns is also important.
}
