% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sankey_functions.R
\name{pathway_to_links}
\alias{pathway_to_links}
\title{A function that converts pathway to links for sankey (Note you need to remove the describers before using this function).}
\usage{
pathway_to_links(pathways, links_as_pos = TRUE, remove_na = T)
}
\arguments{
\item{pathways}{pathways}

\item{links_as_pos}{links_as_pos}

\item{remove_na}{remove_na}
}
\value{
links
}
\description{
A function that converts pathway to links for sankey (Note you need to remove the describers before using this function).
}
