% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sankey_functions.R
\name{sankey_filtered}
\alias{sankey_filtered}
\title{Create sankey diagram of the filtered data.}
\usage{
sankey_filtered(
  pathways,
  flow_columns,
  filters = "",
  union = F,
  fontFamily = NULL,
  fontSize = 7,
  do_SDC = FALSE,
  ...
)
}
\arguments{
\item{pathways}{student pathways with number of students.}

\item{flow_columns}{Column names (or index) of the 'stages' in pathways that we want to create a sankey diagram of.}

\item{filters}{a character vector describing the filters to apply to the data. Of the format `XX: YY` where XX is the stage and YY the node, such as 'FSM: Yes'.}

\item{union}{Flag (TRUE/FALSE) for whether we want the union of filters (only across single stages)}

\item{fontFamily}{font family for the node text labels.}

\item{fontSize}{numeric font size in pixels for the node text labels.}

\item{do_SDC}{Flag (TRUE/FALSE) for whether you want to do SDC via the `SDC_to_links()` function.}

\item{...}{arguments to `SDC_to_links()` or `networkD3::sankeyNetwork()`.}
}
\value{
sankey diagram
}
\description{
Create sankey diagram of the filtered data.
}
