% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sankey_functions.R
\name{sankey_trace}
\alias{sankey_trace}
\title{A function that returns a sankey 'traced' diagram.}
\usage{
sankey_trace(
  pathways,
  flow_columns,
  split_by,
  union = F,
  colors = c("red", "gray", "antiquewhite"),
  FontFamily = NULL,
  FontSize = 7
)
}
\arguments{
\item{pathways}{pathways}

\item{flow_columns}{flow_columns}

\item{split_by}{split_by}

\item{union}{union}

\item{colors}{colors}

\item{FontFamily}{colors}

\item{FontSize}{colors}
}
\value{
traced sankey diagram
}
\description{
A function that returns a sankey 'traced' diagram.
}
